Pull out common asy functions
