+++ mathics/builtin/drawing/graphics3d.py
@@ -4,9 +4,11 @@
 Three-Dimensional Graphics
 """
 
-
-import numbers
 from mathics.version import __version__  # noqa used in loading to check consistency.
+
+import html
+import json
+import numbers
 
 from mathics.core.expression import (
     Expression,
@@ -15,25 +17,23 @@ from mathics.core.expression import (
     SymbolList,
 )
 from mathics.core.formatter import lookup_method
+
 from mathics.builtin.base import BoxConstructError, Builtin, InstanceableBuiltin
 from mathics.builtin.graphics import (
     Graphics,
     GraphicsBox,
     PolygonBox,
-    create_pens,
     _Color,
     LineBox,
     PointBox,
     Style,
     RGBColor,
     get_class,
-    asy_number,
     CoordinatesError,
     _GraphicsElements,
 )
 
-import json
-import html
+from mathics.formatter.asy_fns import asy_create_pens, asy_number
 
 
 def coords3D(value):
@@ -445,7 +445,6 @@ class Graphics3DBox(GraphicsBox):
 
         elements._apply_boxscaling(boxscale)
 
-
         format_fn = lookup_method(elements, "asy")
         if format_fn is not None:
             asy = format_fn(elements)
@@ -473,11 +472,11 @@ class Graphics3DBox(GraphicsBox):
 
         for i, line in enumerate(boundbox_lines):
             if i in axes_indices:
-                pen = create_pens(
+                pen = asy_create_pens(
                     edge_color=RGBColor(components=(0, 0, 0, 1)), stroke_width=1.5
                 )
             else:
-                pen = create_pens(
+                pen = asy_create_pens(
                     edge_color=RGBColor(components=(0.4, 0.4, 0.4, 1)), stroke_width=1
                 )
 
@@ -490,7 +489,7 @@ class Graphics3DBox(GraphicsBox):
 
         # Draw axes ticks
         ticklength = 0.05 * max([xmax - xmin, ymax - ymin, zmax - zmin])
-        pen = create_pens(
+        pen = asy_create_pens(
             edge_color=RGBColor(components=(0, 0, 0, 1)), stroke_width=1.2
         )
         for xi in axes_indices:
@@ -825,22 +824,6 @@ class Point3DBox(PointBox):
             )
         return data
 
-    def to_asy(self):
-        face_color = self.face_color
-
-        # Tempoary bug fix: default Point color should be black not white
-        if list(face_color.to_rgba()[:3]) == [1, 1, 1]:
-            face_color = RGBColor(components=(0, 0, 0, face_color.to_rgba()[3]))
-
-        pen = create_pens(face_color=face_color, is_face_element=False)
-
-        return "".join(
-            "path3 g={0}--cycle;dot(g, {1});".format(
-                "--".join("(%.5g,%.5g,%.5g)" % coords.pos()[0] for coords in line), pen
-            )
-            for line in self.lines
-        )
-
     def extent(self):
         result = []
         for line in self.lines:
@@ -1114,21 +1097,6 @@ class Sphere3DBox(_Graphics3DElement):
 
         self.points = [Coords3D(graphics, pos=point) for point in points]
         self.radius = item.leaves[1].to_python()
-
-    def to_asy(self):
-        # l = self.style.get_line_width(face_element=True)
-
-        if self.face_color is None:
-            face_color = (1, 1, 1)
-        else:
-            face_color = self.face_color.to_js()
-
-        return "".join(
-            "draw(surface(sphere({0}, {1})), rgb({2},{3},{4}));".format(
-                tuple(coord.pos()[0]), self.radius, *face_color[:3]
-            )
-            for coord in self.points
-        )
 
     def to_json(self):
         face_color = self.face_color
