+++ mathics/formatter/asy.py
@@ -4,15 +4,10 @@
 Format a Mathics object as an Aymptote string
 """
 
-from itertools import chain
-
 from mathics.builtin.graphics import (
-    _color,
     _Color,
     _ArcBox,
-    ArrowBox,    # FIXME move elsewhere
-    asy_number,  # FIXME move elsewhere
-    create_pens,
+    ArrowBox,
     BezierCurveBox,
     FilledCurveBox,
     GraphicsBox,
@@ -35,6 +30,7 @@ from mathics.builtin.drawing.graphics3d import (
 )
 
 from mathics.core.formatter import lookup_method, add_conversion_fn
+from mathics.formatter.asy_fns import asy_bezier, asy_color as _color, asy_create_pens, asy_number
 
 
 class _ASYTransform:
@@ -72,78 +68,12 @@ class _ASYTransform:
         return self._template % (" * ".join(self.transforms), asy)
 
 
-def _asy_bezier(*segments):
-    # see http://asymptote.sourceforge.net/doc/Bezier-curves.html#Bezier-curves
-
-    while segments and not segments[0][1]:
-        segments = segments[1:]
-
-    if not segments:
-        return
-
-    def cubic(p0, p1, p2, p3):
-        return "..controls(%.5g,%.5g) and (%.5g,%.5g)..(%.5g,%.5g)" % tuple(
-            list(chain(p1, p2, p3))
-        )
-
-    def quadratric(qp0, qp1, qp2):
-        # asymptote only supports cubic beziers, so we convert this quadratic
-        # bezier to a cubic bezier, see http://fontforge.github.io/bezier.html
-
-        # CP0 = QP0
-        # CP3 = QP2
-        # CP1 = QP0 + 2 / 3 * (QP1 - QP0)
-        # CP2 = QP2 + 2 / 3 * (QP1 - QP2)
-
-        qp0x, qp0y = qp0
-        qp1x, qp1y = qp1
-        qp2x, qp2y = qp2
-
-        t = 2.0 / 3.0
-        cp0 = qp0
-        cp1 = (qp0x + t * (qp1x - qp0x), qp0y + t * (qp1y - qp0y))
-        cp2 = (qp2x + t * (qp1x - qp2x), qp2y + t * (qp1y - qp2y))
-        cp3 = qp2
-
-        return cubic(cp0, cp1, cp2, cp3)
-
-    def linear(p0, p1):
-        return "--(%.5g,%.5g)" % p1
-
-    forms = (linear, quadratric, cubic)
-
-    def path(max_degree, p):
-        max_degree = min(max_degree, len(forms))
-        while p:
-            n = min(max_degree, len(p) - 1)  # 1, 2, or 3
-            if n < 1:
-                break
-            yield forms[n - 1](*p[: n + 1])
-            p = p[n:]
-
-    k, p = segments[0]
-    yield "(%.5g,%.5g)" % p[0]
-
-    connect = []
-    for k, p in segments:
-        for s in path(k, list(chain(connect, p))):
-            yield s
-        connect = p[-1:]
-
-
-# def asy_number(value):
-#     return "%.5g" % value
-
-
-add_conversion_fn(_Color)
-
-
 def arrowbox(self) -> str:
     width = self.style.get_line_width(face_element=False)
-    pen = create_pens(edge_color=self.edge_color, stroke_width=width)
+    pen = asy_create_pens(edge_color=self.edge_color, stroke_width=width)
     polyline = self.curve.make_draw_asy(pen)
 
-    arrow_pen = create_pens(face_color=self.edge_color, stroke_width=width)
+    arrow_pen = asy_create_pens(face_color=self.edge_color, stroke_width=width)
 
     def polygon(points):
         yield "filldraw("
@@ -161,11 +91,11 @@ add_conversion_fn(ArrowBox)
 
 def beziercurvebox(self) -> str:
     line_width = self.style.get_line_width(face_element=False)
-    pen = create_pens(edge_color=self.edge_color, stroke_width=line_width)
+    pen = asy_create_pens(edge_color=self.edge_color, stroke_width=line_width)
 
     asy = ""
     for line in self.lines:
-        for path in _asy_bezier((self.spline_degree, [xy.pos() for xy in line])):
+        for path in asy_bezier((self.spline_degree, [xy.pos() for xy in line])):
             if path[:2] == "..":
                 path = "(0.,0.)" + path
             asy += "draw(%s, %s);" % (path, pen)
@@ -181,7 +111,7 @@ add_conversion_fn(BezierCurveBox)
 
 def filledcurvebox(self) -> str:
     line_width = self.style.get_line_width(face_element=False)
-    pen = create_pens(edge_color=self.edge_color, stroke_width=line_width)
+    pen = asy_create_pens(edge_color=self.edge_color, stroke_width=line_width)
 
     if not pen:
         pen = "currentpen"
@@ -189,7 +119,7 @@ def filledcurvebox(self) -> str:
     def components():
         for component in self.components:
             transformed = [(k, [xy.pos() for xy in p]) for k, p in component]
-            yield "fill(%s--cycle, %s);" % ("".join(_asy_bezier(*transformed)), pen)
+            yield "fill(%s--cycle, %s);" % ("".join(asy_bezier(*transformed)), pen)
 
     return "".join(components())
 
@@ -269,10 +199,11 @@ graphics3delements = graphicselements
 
 add_conversion_fn(Graphics3DElements)
 
+
 def insetbox(self) -> str:
     x, y = self.pos.pos()
     content = self.content.boxes_to_tex(evaluation=self.graphics.evaluation)
-    pen = create_pens(edge_color=self.color)
+    pen = asy_create_pens(edge_color=self.color)
     asy = 'label("$%s$", (%s,%s), (%s,%s), %s);' % (
         content,
         x,
@@ -286,9 +217,10 @@ def insetbox(self) -> str:
 
 add_conversion_fn(InsetBox)
 
+
 def line3dbox(self):
     # l = self.style.get_line_width(face_element=False)
-    pen = create_pens(edge_color=self.edge_color, stroke_width=1)
+    pen = asy_create_pens(edge_color=self.edge_color, stroke_width=1)
 
     return "".join(
         "draw({0}, {1});".format(
@@ -298,12 +230,13 @@ def line3dbox(self):
         for line in self.lines
     )
 
+
 add_conversion_fn(Line3DBox)
 
 
 def linebox(self) -> str:
     line_width = self.style.get_line_width(face_element=False)
-    pen = create_pens(edge_color=self.edge_color, stroke_width=line_width)
+    pen = asy_create_pens(edge_color=self.edge_color, stroke_width=line_width)
     asy = ""
     for line in self.lines:
         path = "--".join(["(%.5g,%5g)" % coords.pos() for coords in line])
@@ -322,7 +255,7 @@ def point3dbox(self) -> str:
     if list(face_color.to_rgba()[:3]) == [1, 1, 1]:
         face_color = RGBColor(components=(0, 0, 0, face_color.to_rgba()[3]))
 
-    pen = create_pens(face_color=face_color, is_face_element=False)
+    pen = asy_create_pens(face_color=face_color, is_face_element=False)
 
     return "".join(
         "path3 g={0}--cycle;dot(g, {1});".format(
@@ -331,10 +264,12 @@ def point3dbox(self) -> str:
         for line in self.lines
     )
 
+
 add_conversion_fn(Point3DBox)
 
+
 def pointbox(self) -> str:
-    pen = create_pens(face_color=self.face_color, is_face_element=False)
+    pen = asy_create_pens(face_color=self.face_color, is_face_element=False)
 
     asy = ""
     for line in self.lines:
@@ -347,13 +282,14 @@ def pointbox(self) -> str:
 
 add_conversion_fn(PointBox)
 
+
 def polygon3dbox(self) -> str:
     l = self.style.get_line_width(face_element=True)
     if self.vertex_colors is None:
         face_color = self.face_color
     else:
         face_color = None
-    pen = create_pens(
+    pen = asy_create_pens(
         edge_color=self.edge_color,
         face_color=face_color,
         stroke_width=l,
@@ -372,7 +308,9 @@ def polygon3dbox(self) -> str:
     # print("### polygon3dbox", asy)
     return asy
 
+
 add_conversion_fn(Polygon3DBox)
+
 
 def polygonbox(self):
     line_width = self.style.get_line_width(face_element=True)
@@ -380,7 +318,7 @@ def polygonbox(self):
         face_color = self.face_color
     else:
         face_color = None
-    pens = create_pens(
+    pens = asy_create_pens(
         edge_color=self.edge_color,
         face_color=face_color,
         stroke_width=line_width,
@@ -426,7 +364,9 @@ def rectanglebox(self) -> str:
     line_width = self.style.get_line_width(face_element=True)
     x1, y1 = self.p1.pos()
     x2, y2 = self.p2.pos()
-    pens = create_pens(self.edge_color, self.face_color, line_width, is_face_element=True)
+    pens = asy_create_pens(
+        self.edge_color, self.face_color, line_width, is_face_element=True
+    )
     x1, x2, y1, y2 = asy_number(x1), asy_number(x2), asy_number(y1), asy_number(y2)
     return "filldraw((%s,%s)--(%s,%s)--(%s,%s)--(%s,%s)--cycle, %s);" % (
         x1,
@@ -450,7 +390,7 @@ def _roundbox(self):
     rx -= x
     ry -= y
     line_width = self.style.get_line_width(face_element=self.face_element)
-    pen = create_pens(
+    pen = asy_create_pens(
         edge_color=self.edge_color,
         face_color=self.face_color,
         stroke_width=line_width,
@@ -469,7 +409,8 @@ def _roundbox(self):
 
 add_conversion_fn(_RoundBox)
 
-def sphere3dbox(self):
+
+def sphere3dbox(self) -> str:
     # l = self.style.get_line_width(face_element=True)
 
     if self.face_color is None:
@@ -483,5 +424,6 @@ def sphere3dbox(self):
         )
         for coord in self.points
     )
+
 
 add_conversion_fn(Sphere3DBox)
